move state class to separate file
